move get_scalar_element to its own header, add support for (eventual) libcudf list_scalar
