+++ modules/cudf/src/node_cudf/utilities/cpp_to_napi.hpp
@@ -14,6 +14,8 @@
 
 #pragma once
 
+#include "scalar_to_value.hpp"
+
 #include <nv_node/utilities/cpp_to_napi.hpp>
 
 #include <cudf/scalar/scalar.hpp>
@@ -82,87 +84,6 @@ inline Napi::Value CPPToNapi::operator()(cudf::timestamp_ns const& val) const {
   return (*this)(val.time_since_epoch());
 }
 
-namespace detail {
-
-struct get_scalar_value {
-  Napi::Env env;
-  template <typename T>
-  inline std::enable_if_t<cudf::is_index_type<T>(), Napi::Value> operator()(
-    std::unique_ptr<cudf::scalar> const& scalar, cudaStream_t stream = 0) {
-    if (!scalar->is_valid(stream)) { return env.Null(); }
-    switch (scalar->type().id()) {
-      case cudf::type_id::INT64:
-        return Napi::BigInt::New(
-          env, static_cast<cudf::numeric_scalar<int64_t>*>(scalar.get())->value(stream));
-      case cudf::type_id::UINT64:
-        return Napi::BigInt::New(
-          env, static_cast<cudf::numeric_scalar<uint64_t>*>(scalar.get())->value(stream));
-      default:
-        return Napi::Number::New(
-          env, static_cast<cudf::numeric_scalar<T>*>(scalar.get())->value(stream));
-    }
-  }
-  template <typename T>
-  inline std::enable_if_t<cudf::is_floating_point<T>(), Napi::Value> operator()(
-    std::unique_ptr<cudf::scalar> const& scalar, cudaStream_t stream = 0) {
-    return scalar->is_valid(stream)
-             ? Napi::Number::New(env,
-                                 static_cast<cudf::numeric_scalar<T>*>(scalar.get())->value(stream))
-             : env.Null();
-  }
-  template <typename T>
-  inline std::enable_if_t<std::is_same<T, bool>::value, Napi::Value> operator()(
-    std::unique_ptr<cudf::scalar> const& scalar, cudaStream_t stream = 0) {
-    return scalar->is_valid(stream)
-             ? Napi::Boolean::New(
-                 env, static_cast<cudf::numeric_scalar<bool>*>(scalar.get())->value(stream))
-             : env.Null();
-  }
-  template <typename T>
-  inline std::enable_if_t<std::is_same<T, cudf::string_view>::value, Napi::Value> operator()(
-    std::unique_ptr<cudf::scalar> const& scalar, cudaStream_t stream = 0) {
-    return scalar->is_valid(stream)
-             ? CPPToNapi(env)(static_cast<cudf::string_scalar*>(scalar.get())->to_string(stream))
-             : env.Null();
-  }
-  template <typename T>
-  inline std::enable_if_t<cudf::is_duration<T>(), Napi::Value> operator()(
-    std::unique_ptr<cudf::scalar> const& scalar, cudaStream_t stream = 0) {
-    return scalar->is_valid(stream)
-             ? CPPToNapi(env)(static_cast<cudf::duration_scalar<T>*>(scalar.get())->value(stream))
-             : env.Null();
-  }
-  template <typename T>
-  inline std::enable_if_t<cudf::is_timestamp<T>(), Napi::Value> operator()(
-    std::unique_ptr<cudf::scalar> const& scalar, cudaStream_t stream = 0) {
-    return scalar->is_valid(stream)
-             ? CPPToNapi(env)(static_cast<cudf::timestamp_scalar<T>*>(scalar.get())->value(stream))
-             : env.Null();
-  }
-  template <typename T>
-  inline std::enable_if_t<cudf::is_fixed_point<T>(), Napi::Value> operator()(
-    std::unique_ptr<cudf::scalar> const& scalar, cudaStream_t stream = 0) {
-    return scalar->is_valid(stream)
-             ? CPPToNapi(env)(
-                 static_cast<cudf::fixed_point_scalar<T>*>(scalar.get())->value(stream))
-             : env.Null();
-  }
-  template <typename T>
-  inline std::enable_if_t<!(cudf::is_index_type<T>() ||                   //
-                            cudf::is_floating_point<T>() ||               //
-                            std::is_same<T, bool>::value ||               //
-                            std::is_same<T, cudf::string_view>::value ||  //
-                            cudf::is_duration<T>() ||                     //
-                            cudf::is_timestamp<T>() ||                    //
-                            cudf::is_fixed_point<T>()),
-                          Napi::Value>
-  operator()(std::unique_ptr<cudf::scalar> const& scalar, cudaStream_t stream = 0) {
-    NAPI_THROW(Napi::Error::New(env, "Unsupported dtype"));
-  }
-};
-
-}  // namespace detail
-
 template <>
 inline Napi::Value CPPToNapi::operator()(std::unique_ptr<cudf::scalar> const& scalar) const {
   return cudf::type_dispatcher(scalar->type(), detail::get_scalar_value{Env()}, scalar);
@@ -225,11 +146,6 @@ inline Value Value::From(napi_env env, cudf::timestamp_us const& val) {
 template <>
 inline Value Value::From(napi_env env, cudf::timestamp_ns const& val) {
   return Value::From(env, val.time_since_epoch());
-}
-
-template <>
-inline Value Value::From(napi_env env, std::unique_ptr<cudf::scalar> const& scalar) {
-  return cudf::type_dispatcher(scalar->type(), nv::detail::get_scalar_value{env}, scalar);
 }
 
 }  // namespace Napi
